Xbox now has its own version of the Song History page. Various attempts to speed loading up.
