+++ src/Neptunium/Neptunium.csproj
@@ -239,6 +239,9 @@
     <Compile Include="View\XboxShellView.xaml.cs">
       <DependentUpon>XboxShellView.xaml</DependentUpon>
     </Compile>
+    <Compile Include="View\XboxSongHistoryPage.xaml.cs">
+      <DependentUpon>XboxSongHistoryPage.xaml</DependentUpon>
+    </Compile>
     <Compile Include="View\XboxStationsPage.xaml.cs">
       <DependentUpon>XboxStationsPage.xaml</DependentUpon>
     </Compile>
@@ -420,6 +423,10 @@
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Include="View\XboxShellView.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="View\XboxSongHistoryPage.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
