+++ src/Neptunium/View/SongHistoryPage.xaml
@@ -14,43 +14,6 @@
 
     <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
         <Grid.Resources>
-            <Style x:Key="TitleSafeListViewStyle" TargetType="ListView">
-                <!-- modified version of the style from here: https://docs.microsoft.com/en-us/windows/uwp/design/devices/designing-for-tv#tv-safe-area -->
-                <Setter Property="Template">
-                    <Setter.Value>
-                        <ControlTemplate TargetType="ListView">
-                            <Border BorderBrush="{TemplateBinding BorderBrush}"
-                                        Background="{TemplateBinding Background}"
-                                        BorderThickness="{TemplateBinding BorderThickness}">
-                                <ScrollViewer x:Name="ScrollViewer"
-                                                  TabNavigation="{TemplateBinding TabNavigation}"
-                                                  HorizontalScrollMode="{TemplateBinding ScrollViewer.HorizontalScrollMode}"
-                                                  HorizontalScrollBarVisibility="{TemplateBinding ScrollViewer.HorizontalScrollBarVisibility}"
-                                                  IsHorizontalScrollChainingEnabled="{TemplateBinding ScrollViewer.IsHorizontalScrollChainingEnabled}"
-                                                  VerticalScrollMode="{TemplateBinding ScrollViewer.VerticalScrollMode}"
-                                                  VerticalScrollBarVisibility="{TemplateBinding ScrollViewer.VerticalScrollBarVisibility}"
-                                                  IsVerticalScrollChainingEnabled="{TemplateBinding ScrollViewer.IsVerticalScrollChainingEnabled}"
-                                                  IsHorizontalRailEnabled="{TemplateBinding ScrollViewer.IsHorizontalRailEnabled}"
-                                                  IsVerticalRailEnabled="{TemplateBinding ScrollViewer.IsVerticalRailEnabled}"
-                                                  ZoomMode="{TemplateBinding ScrollViewer.ZoomMode}"
-                                                  IsDeferredScrollingEnabled="{TemplateBinding ScrollViewer.IsDeferredScrollingEnabled}"
-                                                  BringIntoViewOnFocusChange="{TemplateBinding ScrollViewer.BringIntoViewOnFocusChange}"
-                                  AutomationProperties.AccessibilityView="Raw">
-                                    <ItemsPresenter Header="{TemplateBinding Header}"
-                                                        HeaderTemplate="{TemplateBinding HeaderTemplate}"
-                                                        HeaderTransitions="{TemplateBinding HeaderTransitions}"
-                                                        Footer="{TemplateBinding Footer}"
-                                                        FooterTemplate="{TemplateBinding FooterTemplate}"
-                                                        FooterTransitions="{TemplateBinding FooterTransitions}"
-                                                        Padding="{TemplateBinding Padding}"
-                                                        Margin="0,0,0,27"/>
-                                </ScrollViewer>
-                            </Border>
-                        </ControlTemplate>
-                    </Setter.Value>
-                </Setter>
-            </Style>
-
             <!--<CollectionViewSource x:Key="GroupedSongItems" IsSourceGrouped="True" Source="{Binding History, UpdateSourceTrigger=PropertyChanged}" />-->
             <cconv:RelativeTimeConverter TimeMode="Past" x:Key="TimeConv" HourFormatString="Today" MinuteFormatString="Today" SecondFormatString="Today" />
 
@@ -157,9 +120,9 @@
         </Grid.Resources>
 
         <ListView x:Name="SongHistoryListView"
-                ScrollViewer.IsVerticalScrollChainingEnabled="True"
-                ItemsSource="{Binding History, UpdateSourceTrigger=PropertyChanged}"
-                Style="{StaticResource TitleSafeListViewStyle}">
+                  ScrollViewer.IsVerticalScrollChainingEnabled="True"
+                  ItemTemplate="{StaticResource MobileSongHistoryItemTemplate}"
+                  ItemsSource="{Binding History, UpdateSourceTrigger=PropertyChanged}">
             <ListView.HeaderTemplate>
                 <DataTemplate>
                     <TextBlock Text="Play History"
@@ -174,17 +137,13 @@
             </ListView.Transitions>
         </ListView>
 
+        <ProgressRing VerticalAlignment="Center" HorizontalAlignment="Center"
+                      IsActive="{Binding IsBusy, UpdateSourceTrigger=PropertyChanged}"
+                      Height="60"
+                      Width="60" />
+
         <VisualStateManager.VisualStateGroups>
             <VisualStateGroup>
-                <VisualState x:Name="XboxVisualState">
-                    <VisualState.StateTriggers>
-                        <wst:DeviceFamilyStateTrigger DeviceFamily="Xbox" />
-                    </VisualState.StateTriggers>
-                    <VisualState.Setters>
-                        <Setter Target="SongHistoryListView.Margin" Value="48 0 48 0" />
-                        <Setter Target="SongHistoryListView.ItemTemplate" Value="{StaticResource TabletOrHigherSongHistoryItemTemplate}" />
-                    </VisualState.Setters>
-                </VisualState>
                 <VisualState x:Name="TabletOrHigherVisualState">
                     <VisualState.StateTriggers>
                         <wst:CompositeStateTrigger Operator="And">
